risk-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pydantic==2.5.0
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
PyJWT==2.8.0
